Refactoring tracing interfaces to `camel-api` and `camel-support`
This allows moving the span decorators closer to the components.
