+++ components/camel-tracing/src/main/java/org/apache/camel/tracing/decorators/AbstractMessagingSpanDecorator.java
@@ -20,13 +20,14 @@ import java.util.*;
 
 import org.apache.camel.Endpoint;
 import org.apache.camel.Exchange;
+import org.apache.camel.support.tracing.decorators.AbstractSpanDecorator;
+import org.apache.camel.support.tracing.propagators.CamelMessagingHeadersExtractAdapter;
+import org.apache.camel.support.tracing.propagators.CamelMessagingHeadersInjectAdapter;
 import org.apache.camel.tracing.ExtractAdapter;
 import org.apache.camel.tracing.InjectAdapter;
 import org.apache.camel.tracing.SpanAdapter;
 import org.apache.camel.tracing.SpanKind;
 import org.apache.camel.tracing.TagConstants;
-import org.apache.camel.tracing.propagation.CamelMessagingHeadersExtractAdapter;
-import org.apache.camel.tracing.propagation.CamelMessagingHeadersInjectAdapter;
 
 public abstract class AbstractMessagingSpanDecorator extends AbstractSpanDecorator {
