+++ components/camel-tracing/src/main/java/org/apache/camel/tracing/decorators/SedaSpanDecorator.java
@@ -16,6 +16,8 @@
  */
 package org.apache.camel.tracing.decorators;
 
+import org.apache.camel.support.tracing.decorators.AbstractInternalSpanDecorator;
+
 public class SedaSpanDecorator extends AbstractInternalSpanDecorator {
 
     @Override
