+++ components/camel-tracing/src/main/java/org/apache/camel/tracing/decorators/TimerSpanDecorator.java
@@ -18,6 +18,7 @@ package org.apache.camel.tracing.decorators;
 
 import org.apache.camel.Endpoint;
 import org.apache.camel.Exchange;
+import org.apache.camel.support.tracing.decorators.AbstractSpanDecorator;
 
 public class TimerSpanDecorator extends AbstractSpanDecorator {
