+++ components/camel-tracing/src/test/java/org/apache/camel/tracing/decorators/RestSpanDecoratorTest.java
@@ -21,7 +21,7 @@ import java.util.Arrays;
 import org.apache.camel.Endpoint;
 import org.apache.camel.Exchange;
 import org.apache.camel.Message;
-import org.apache.camel.tracing.MockSpanAdapter;
+import org.apache.camel.support.tracing.decorators.InMemorySpanAdapter;
 import org.apache.camel.tracing.SpanDecorator;
 import org.junit.jupiter.api.Test;
 import org.mockito.Mockito;
@@ -84,7 +84,7 @@ public class RestSpanDecoratorTest {
 
         SpanDecorator decorator = new RestSpanDecorator();
 
-        MockSpanAdapter span = new MockSpanAdapter();
+        InMemorySpanAdapter span = new InMemorySpanAdapter();
 
         decorator.pre(span, exchange, endpoint);
