+++ core/camel-api/src/main/java/org/apache/camel/tracing/TagConstants.java
