+++ core/camel-support/src/main/java/org/apache/camel/support/tracing/propagators/CamelHeadersInjectAdapter.java
@@ -14,16 +14,24 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.camel.tracing;
+package org.apache.camel.support.tracing.propagators;
 
-import java.util.Iterator;
 import java.util.Map;
-import java.util.Set;
 
-public interface ExtractAdapter {
-    Iterator<Map.Entry<String, Object>> iterator();
+import org.apache.camel.tracing.InjectAdapter;
 
-    Object get(String key);
+public final class CamelHeadersInjectAdapter implements InjectAdapter {
+    private final Map<String, Object> map;
 
-    Set<String> keys();
+    public CamelHeadersInjectAdapter(final Map<String, Object> map) {
+        this.map = map;
+    }
+
+    @Override
+    public void put(String key, String value) {
+        // Assume any header property that begins with 'Camel' is for internal use
+        if (!key.startsWith("Camel")) {
+            this.map.put(key, value);
+        }
+    }
 }
