+++ core/camel-support/src/main/java/org/apache/camel/support/tracing/propagators/CamelMessagingHeadersInjectAdapter.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.camel.tracing.propagation;
+package org.apache.camel.support.tracing.propagators;
 
 import java.util.Map;
