+++ core/camel-support/src/test/java/org/apache/camel/support/tracing/propagators/CamelMessagingHeadersExtractAdapterTest.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.camel.tracing.propagation;
+package org.apache.camel.support.tracing.propagators;
 
 import java.util.HashMap;
 import java.util.Iterator;
@@ -23,7 +23,7 @@ import java.util.Map;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
-import static org.apache.camel.tracing.propagation.CamelMessagingHeadersInjectAdapter.JMS_DASH;
+import static org.apache.camel.support.tracing.propagators.CamelMessagingHeadersInjectAdapter.JMS_DASH;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
